oa_server/requirements.txt
Django
pylint-django
pylint-exit
djangorestframework
djangorestframework-csv
django-cors-headers
psycopg2-binary
coverage